site/flask/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy<2.0
psycopg2-binary==2.9.9
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.1